// src/graphql/generated/types.tsx
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  Upload: any;
};
// ...
export type DaylyParams = {
  __typename?: "DaylyParams";
  dayOfWeeks: Array<DaysOfWeek>;
  everyWeek?: Maybe<Scalars["Float"]>;
};
// ...
export enum EventType {
  AdminSurvey = "ADMIN_SURVEY",
  /** APPOINTMENT */
  Appointment = "APPOINTMENT",
  /** CALLREMINDER */
  Callreminder = "CALLREMINDER",
  /** REMINDER */
  Reminder = "REMINDER",
  /** Record_Reminder */
  RecordReminder = "Record_Reminder",
  /** SCHEDULE */
  Schedule = "SCHEDULE",
}
// ...
export type MonthlyParamsForReminder = {
  __typename?: "MonthlyParamsForReminder";
  months?: Maybe<Array<Months>>;
  onDay?: Maybe<Scalars["Float"]>;
  onWeek?: Maybe<DaylyParams>;
  twicePerMonth?: Maybe<Scalars["Boolean"]>;
};
// ...
export enum RecurrentTypes {
  /** ANNUALLY */
  Annually = "ANNUALLY",
  /** ANYTIME */
  Anytime = "ANYTIME",
  /** DAILY */
  Daily = "DAILY",
  /** HALFYEARLY */
  Halfyearly = "HALFYEARLY",
  /** MONTHLY */
  Monthly = "MONTHLY",
  /** ONCE */
  Once = "ONCE",
  /** PERIODICAL */
  Periodical = "PERIODICAL",
  /** QUARTERLY */
  Quarterly = "QUARTERLY",
  /** WEEKLY */
  Weekly = "WEEKLY",
}
// ...
export enum SurveyEventType {
  /** CALENDAR */
  Calendar = "CALENDAR",
  /** CALL */
  Call = "CALL",
  /** CHAT */
  Chat = "CHAT",
  Other = "OTHER",
  /** TASK */
  Task = "TASK",
}
// ...
export type AdminSurvey = {
  __typename?: "adminSurvey";
  Module?: Maybe<SurveyEventType>;
  _id?: Maybe<Scalars["String"]>;
  allOccurrence?: Maybe<Scalars["Boolean"]>;
  date?: Maybe<Scalars["String"]>;
  daylyParams?: Maybe<DaylyParams>;
  endDate?: Maybe<Scalars["String"]>;
  frequency?: Maybe<Scalars["Float"]>;
  monthlyParams?: Maybe<MonthlyParamsForReminder>;
  parent_id?: Maybe<Scalars["String"]>;
  publishType: SurveyScenarioType;
  recursive?: Maybe<RecurrentTypes>;
  scenario?: Maybe<Array<Scalars["String"]>>;
  startDate?: Maybe<Scalars["String"]>;
  surveyActiveStatus: Scalars["Boolean"];
  thisOccurrence?: Maybe<Scalars["Boolean"]>;
  userId?: Maybe<Scalars["String"]>;
};
// ...
export enum ParticipantAcceptStatus {
  /** ACCEPT */
  Accept = "ACCEPT",
  /** PAUSE */
  Pause = "PAUSE",
  /** PENDING */
  Pending = "PENDING",
  /** REJECT */
  Reject = "REJECT",
}
// ...
export enum SurveyScenarioType {
  /** Draft scenario */
  Draft = "DRAFT",
  /** Master scenario */
  Master = "MASTER",
  /** PAUSE scenario */
  Pause = "PAUSE",
  /** Published scenario */
  Published = "PUBLISHED",
  /** Template scenario */
  Template = "TEMPLATE",
}